fastboot: fix wrong type casting
This patch fixes wrong type casting
to prevent build warning message.

-----------------------------------------------------
drivers/usb/gadget/f_fastboot.c:322:13:
warning: cast from pointer to integer of different size [-Wpointer-to-int-cast]

+++ drivers/usb/gadget/f_fastboot.c
@@ -319,8 +319,7 @@ static int mmc_part_write(struct fastboot_part *fpart, void *buf,
 
 	if (fpart->fs_type & FASTBOOT_FS_EXT4) {
 		p = sprintf(cmd, "ext4_img_write %d %p %d %x",
-			    dev, (unsigned int)buf, fpart->part_num,
-			    (unsigned int)length);
+			    dev, buf, fpart->part_num, (unsigned int)length);
 		ret = run_command(cmd, 0);
 		if (ret < 0)
 			printf("Flash : %s\n", (ret < 0) ? "FAIL" : "DONE");
